pyproject.toml
[tool.poetry]
name = "zadanie-rekrutacyjne-python-sta-"
version = "0.1.0"
description = ""
authors = ["MaximPyanin <maximpyanin@icloud.com>"]
readme = "README.md"
packages = [{include = "articles_scrape"}]

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.11.2"
selenium = "^4.25.0"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
